Flix 43  schedule

+++ calendar-service/src/main/java/com/ovengers/calendarservice/controller/CalendarController.java
@@ -6,10 +6,12 @@ import com.ovengers.calendarservice.dto.response.ScheduleResponseDto;
 import com.ovengers.calendarservice.entity.Schedule;
 import com.ovengers.calendarservice.service.CalendarService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.cglib.core.Local;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDateTime;
 import java.util.*;
@@ -17,17 +19,22 @@ import java.util.*;
 @RestController
 @RequestMapping("/api/schedules")
 @RequiredArgsConstructor
+@Slf4j
 public class CalendarController {
 
     private final CalendarService calendarService;
 
     // 전체 일정 조회
     @GetMapping("")
-    public ResponseEntity<List<ScheduleResponseDto>> getAllSchedules(
-            @RequestParam("start")LocalDateTime start,
-            @RequestParam("end")LocalDateTime end) {
-        List<ScheduleResponseDto> schedules = calendarService.getAllSchedules();
-        return ResponseEntity.ok(schedules);
+    public ResponseEntity<List<ScheduleResponseDto>> getAllSchedules() {
+        try {
+            List<ScheduleResponseDto> schedules = calendarService.getAllSchedules();
+            return ResponseEntity.ok(schedules);
+        } catch (Exception e) {
+            log.error("Error fetching schedules", e); // 예외 로그 기록
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body(Collections.emptyList());
+        }
     }
 
     // 특정 일정 조회
@@ -41,6 +48,7 @@ public class CalendarController {
     // 일정 생성
     @PostMapping("/create-schedule")
     public ResponseEntity<ScheduleResponseDto> addSchedule(@RequestBody ScheduleRequestDto scheduleRequestDto) {
+        log.info("/create-schedule: POST!, dto: {}", scheduleRequestDto);
         ScheduleResponseDto createdSchedule = calendarService.createSchedule(scheduleRequestDto);
         return ResponseEntity.status(HttpStatus.CREATED).body(createdSchedule);
     }
@@ -56,14 +64,14 @@ public class CalendarController {
         return ResponseEntity.ok(modifySchedule);
     }
 
-    // 일정 삭제
-    @DeleteMapping("/delete/{id}")
-    public ResponseEntity<CommonResDto> deleteSchedule(@PathVariable("id") UUID scheduleId) {
-        calendarService.deleteSchedule(scheduleId);
+    @DeleteMapping("/delete-schedule")
+    public ResponseEntity<Void> deleteSchedule(@RequestParam UUID scheduleId) {
+        if (scheduleId == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "scheduleId is required");
+        }
 
+        calendarService.deleteSchedule(scheduleId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
-
-
 
 }
